+++ source/error-handler/jest.config.js
@@ -1,0 +1,6 @@
+module.exports = {
+    roots: ['<rootDir>/lib'],
+    testMatch: ['**/*.spec.js'],
+    coveragePathIgnorePatterns: ['<rootDir>/lib/utils.test.js'],
+    coverageReporters: [['lcov', { projectRoot: '../' }], 'text']
+  };